services/daily-cron/requirements.txt
flask
google-api-python-client
google-auth
google-cloud-secret-manager
google-cloud-firestore
google-cloud-storage
openai
